Geolocation (#4)
* testing message system

* disable masstransit

* rabbitmq

* rabbitmq raiz

* messaging full finalized

* unit of work

* starting geo

* datetime adjusts

* geolocation done

+++ Domain/Features/Orders/Enums/OrderStatus.cs
@@ -1,6 +1,6 @@
 namespace Domain.Enums
 {
-    public enum OrderStatus : int
+    public enum OrderStatus : ushort
     {
         PendingPayment = 1,
         ReadyToAccept = 2,
